OAK-8840: mark packages internal - extract mapper class, fix readBlob
